Feat: added test files, added Enum values, changing types
- adding testfiles from an ib project
- adding new enum values
- changing some date types to accept the other format
- adding new Methode FromFile() for easier file reading

+++ IbFlexReader/IbFlexReader.Contracts/Ib/OpenPosition.cs
@@ -58,7 +58,7 @@
         //Note: The reportDate XML attribute may contain either a date or a string, i.e. reportDate="MULTI"
         public string ReportDate { get; set; }
 
-        public int? Position { get; set; }
+        public double? Position { get; set; }
 
         public double? MarkPrice { get; set; }
